:recycle:refactor: refactor posting dialog 컴포넌트, 스타일, 함수 분리 Related to: #78
